app/requirements.txt
numpy==1.26.3
tensorflow==2.15.0
streamlit==1.30.0
google-generativeai==0.7.2
python-dotenv==1.0.1
Pillow==10.4.0
protobuf==4.25.3
